backend/requirements.txt
# Core Flask and Web Framework
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0

# Computer Vision and Image Processing
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0

# HTTP and API
requests==2.31.0
urllib3==2.1.0

# Spotify Integration
spotipy==2.23.0

# Environment and Configuration
python-dotenv==1.1.0

# AI/ML Inference
inference-sdk==0.9.2
roboflow==1.2.16

# Database
qdrant-client==1.7.0
pymongo==4.6.0

# Additional Utilities
python-dateutil==2.8.2
